18.09.20

+++ src/main/java/com/example/demo/service/BoardService.java
@@ -8,7 +8,7 @@ public interface BoardService {
 
     public void register(BoardGallery boardGallery) throws Exception;
     public List<BoardGallery> list() throws Exception;
-    public BoardGallery read(Integer boardNo) throws Exception;
-    public void remove(Integer boardNo) throws Exception;
+    public BoardGallery read(Long boardNo) throws Exception;
+    public void remove(Long boardNo) throws Exception;
     public void modify(BoardGallery boardGallery) throws Exception;
 }
